+++ src/VerisFlow.LayParser.Core/LabwareInfo.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text;
 
-namespace VerisFlow.VenusDeckParser.Core
+namespace VerisFlow.LayParser.Core
 {
     /// <summary>
     /// Represents a 3D vector for TForm data.