+++ src/VerisFlow.LayParser.Core/CHANGELOG.md
@@ -1,6 +1,6 @@
-# Changelog - VerisFlow.VenusDeckParser.Core
+# Changelog - VerisFlow.LayParser.Core
 
-All notable changes to the `VerisFlow.VenusDeckParser.Core` package will be documented in this file.
+All notable changes to the `VerisFlow.LayParser.Core` package will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Initial release of the `VerisFlow.VenusDeckParser.Core` package at version 0.1.0.
+* Initial release of the `VerisFlow.LayParser.Core` package at version 0.1.0.
 
 * Introduced the `DeckLayoutParser` class to systematically parse Hamilton Venus deck layout (`.lay`) files.